% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allComplements}
\alias{allComplements}
\title{Return all groups for an identity class in a Seurat object}
\usage{
allComplements(seuratObj, idClass, ids1)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{idClass}{Identity class.}

\item{ids1}{Selected class groups.}
}
\value{
A character vector list of identity classes.
}
\description{
This function returns all groups for an identity class in a Seurat object.
}
\keyword{internal}
