% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allGroups}
\alias{allGroups}
\title{Return all groups for an identity class in a Seurat object}
\usage{
allGroups(seuratObj, idClass)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{idClass}{Identity class.}
}
\value{
A character vector list of identity classes.
}
\description{
This function returns all groups for an identity class in a Seurat object.
}
\keyword{internal}
