% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_lists.R
\name{buildDEGList}
\alias{buildDEGList}
\title{Find differentiallty expressed genes between identity groups}
\usage{
buildDEGList(
  seuratObj,
  idClass = "seurat_clusters",
  ids1 = allGroups(seuratObj, idClass),
  ids2 = allComplements(seuratObj, idClass, ids1),
  resNames = ids1,
  logFCThr = 0,
  minPct = 0,
  minPctRatio = 0,
  ...
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{idClass}{Identity class.}

\item{ids1}{Selected class groups.}

\item{ids2}{Selected class groups used for comparison. Defaults to the
complements of \code{ids1}.}

\item{resNames}{Names of the output object.}

\item{logFCThr}{Fold change threshold for testing.}

\item{minPct}{The minimum fraction of in-cluster cells in which tested
genes need to be expressed.}

\item{minPctRatio}{The minimum ratio of in-cluster cells over out-cluster
cells in which a retained gene must be expressed.}

\item{...}{Additional arguments passed to \code{Seurat::FindMarkers}.}
}
\value{
A list of marker data frames.
}
\description{
This function finds markers (positive and negative) for groups belonging to
a given Seurat identity class and performs an additional Bonferroni
correction for multiple testing.
}
\details{
The difference between \code{buildDEGList} and
\code{buildMarkerList} is that \code{buildDEGList} returns all markers,
both upregulated and downregulated. In constrast, \code{buildMarkerList}
returns only upregulated or only downregulated markers, depending on the
\code{invert} parameter.
}
