% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_lists.R
\name{buildMarkerListCore}
\alias{buildMarkerListCore}
\title{Find markers for a Seurat identity group}
\usage{
buildMarkerListCore(
  seuratObj,
  idClass = "seurat_clusters",
  id1 = NULL,
  id2 = NULL,
  onlyPos = TRUE,
  logFCThr = 0,
  minPct = 0,
  minPctRatio = 0,
  nTests = 1,
  ...
)
}
\arguments{
\item{seuratObj}{A Seurat object.}

\item{idClass}{Identity class.}

\item{id1}{Selected group.}

\item{id2}{Group selected for comparison.}

\item{onlyPos}{Whether only positive markers should be computed.}

\item{logFCThr}{Fold change threshold for testing.}

\item{minPct}{The minimum fraction of in-cluster cells in which tested
genes need to be expressed.}

\item{minPctRatio}{The minimum ratio of in-cluster cells over out-cluster
cells in which a retained gene must be expressed.}

\item{...}{Additional arguments passed to \code{Seurat::FindMarkers}.}

\item{ids2}{Selected class groups used for comparison. Ignored
if \code{invert} is \code{TRUE}.}
}
\value{
A list of marker data frames.
}
\description{
This function finds markers for a Seurat identity group and performs an
additional Bonferroni correction for multiple testing.
}
\keyword{internal}
