% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R, R/generics.R
\name{dimredNames.default}
\alias{dimredNames.default}
\alias{dimredNames.Seurat}
\alias{dimredNames.SingleCellExperiment}
\alias{dimredNames}
\title{Extract the names of available dimensionality reductions from object}
\usage{
\method{dimredNames}{default}(scObj)

\method{dimredNames}{Seurat}(scObj)

\method{dimredNames}{SingleCellExperiment}(scObj)

dimredNames(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
A character vector.
}
\description{
This function extracts the names of available dimensionality reductions
from a Seurat or SingleCellExperiment object.
}
\examples{
scObj <- withr::with_seed(1, scuttle::mockSCE(ngenes=20000))
scObj <- scuttle::logNormCounts(scObj)
scObj <- scater::runPCA(scObj)
dimredNames(scObj)

}
