% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentFun}
\alias{documentFun}
\title{Internal function used for documenting other functions}
\usage{
documentFun(
  scObj = NULL,
  title = NULL,
  groupBy = NULL,
  dimred = "umap",
  dims = c(1, 2),
  xLab = "x",
  yLab = "y",
  legendTitle = "Legend",
  legendLabs = c("a", "b"),
  legendPos = "right",
  palette = "Spectral",
  alpha = 1,
  pointSize = 0.8,
  legendTitleSize = 10,
  legendTextSize = 10,
  axisTextSize = 12,
  axisTitleSize = 12,
  ...
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{title}{Plot title.}

\item{groupBy}{Grouping variable. Must exist in the metadata/coldata of
the single-cell expression object.}

\item{dimred}{Dimensionality reduction.}

\item{dims}{A numeric vector of size 2 representing the dimensions selected
for the plot.}

\item{xLab}{x axis label.}

\item{yLab}{y axis label.}

\item{legendTitle}{Legend title.}

\item{legendLabs}{Legend labels.}

\item{legendPos}{Legend position.}

\item{palette}{Color palette.}

\item{alpha}{Opaqueness level.}

\item{pointSize}{Point size.}

\item{legendTitleSize}{Legend title size.}

\item{legendTextSize}{Legend text size.}

\item{axisTextSize}{Axis text size.}

\item{axisTitleSize}{Axis title size.}

\item{...}{Additional arguments passed to \code{henna::centerTitle}.}
}
\value{
\code{NULL}. This function is only used internally for
documentation.
}
\description{
This function is used internally to help document other functions.
}
\keyword{internal}
