% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_lists.R
\name{filterMarkerList}
\alias{filterMarkerList}
\title{Filter marker list based on log2 fold-change and pct.1}
\usage{
filterMarkerList(markerList, logFCThr = 0, pct1Thr = 0)
}
\arguments{
\item{markerList}{List of marker data frames.}

\item{logFCThr}{Value of average log2 fold-change above which markers will
be retained.}

\item{pct1Thr}{Value of fraction of cells expressing the marker above which
markers will be retained.}
}
\value{
A list of filtered marker data frames.
}
\description{
This function filters a marker list based on log2 fold-change and pct.1.
}
\keyword{internal}
