% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToChar}
\alias{listToChar}
\title{Join the elements of each vector in a list into a character}
\usage{
listToChar(v, joinChar = ", ")
}
\arguments{
\item{v}{Vector list.}

\item{joinChar}{Character used when joining the elements of each vector in
a list.}
}
\value{
A character vector.
}
\description{
This function joins the elements of each vector in a list into a character.
}
\examples{
v <- list(c(1, 2, 3), c(3, 4, 5), c('apple', 'banana'))
listToChar(v)

}
