% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{markerDFPair}
\alias{markerDFPair}
\title{Extract two marker data frames from two named marker lists}
\usage{
markerDFPair(markerList1, markerList2, name1, name2)
}
\arguments{
\item{markerList1}{A named list of marker data frames.}

\item{markerList2}{A named list of marker data frames.}

\item{name1}{Name of the marker data frame that will be selected from
\code{markerList1}.}

\item{name2}{Name of the marker data frame that will be selected from
\code{markerList2}.}
}
\value{
A named list of two marker data frames.
}
\description{
This function receives two named marker lists as input (they can be
identical) and extracts a specified marker data frame from each. The two
extracted marker data frames are grouped in a named list.
}
