% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadataNames}
\alias{metadataNames}
\alias{metadataNames<-}
\title{Return metadata names}
\usage{
metadataNames(scObj)

metadataNames(scObj) <- value
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{value}{A data frame to replace metadata with.}
}
\value{
The names of the metadata columns.
}
\description{
This function extracts metadata names from a Seurat or
SingleCellExperiment object. It can also be used to modify metadata names.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='scLang')
sceObj <- qs::qread(scePath)
colNames <- metadataNames(sceObj)

}
