% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_utils.R
\name{pickFeature}
\alias{pickFeature}
\title{Extract feature data to be used by \code{featurePlot} or \code{violinPlot}}
\usage{
pickFeature(scObj, feature)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{feature}{A gene name or metadata column name.}
}
\value{
A numeric vector.
}
\description{
This function extracts feature data to be used by \code{featurePlot}
or \code{violinPlot}.
}
\keyword{internal}
