% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_and_percentages.R
\name{scColPairPercs}
\alias{scColPairPercs}
\title{Extract percentages from two columns of a single-cell expression
object}
\usage{
scColPairPercs(scObj, col1, col2, sigDigits = 2)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col1}{Column as string.}

\item{col2}{Column as string.}

\item{sigDigits}{Number of significant digits.}
}
\value{
A data frame listing the percentages of all combinations of pairs
from two categorical columns.
}
\description{
This function extracts percentage information from two columns of a
single-cell expression object. For each i x j combination with i taken
from column 1 and j taken from column 2, the function reports the percentage
that i contributes to all combinations involving j.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='scLang')
sceObj <- qs2::qs_read(scePath)
scColPairPercs(sceObj, 'Cluster', 'Donor')

}
