% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_mat.R, R/generics.R
\name{scDimredMat.default}
\alias{scDimredMat.default}
\alias{scDimredMat.Seurat}
\alias{scDimredMat.SingleCellExperiment}
\alias{scDimredMat}
\title{Extracts a dimensionality reduction matrix from object.}
\usage{
\method{scDimredMat}{default}(scObj, dimred)

\method{scDimredMat}{Seurat}(scObj, dimred)

\method{scDimredMat}{SingleCellExperiment}(scObj, dimred)

scDimredMat(scObj, dimred)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{dimred}{Dimensionality reduction.}
}
\value{
A dimensionality reduction matrix.
}
\description{
This function extracts a dimensionality reduction matrix from a Seurat or
SingleCellExperiment object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='scLang')
sceObj <- qs2::qs_read(scePath)
pcaMat <- scDimredMat(sceObj, 'pca')


}
