% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLang-package.R
\docType{package}
\name{scLang-package}
\alias{scLang}
\alias{scLang-package}
\title{scLang: A unified language for interacting with Seurat and SingleCellExperiment}
\description{
scLang is a suite for package development for scRNA-seq analysis. It offers functions that can operate on both Seurat and SingleCellExperiment objects. These functions are primarily aimed to help developers build tools compatible with both types of input.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrei-stoica26/scLang}
  \item Report bugs at \url{https://github.com/andrei-stoica26/scLang/issues}
}

}
\author{
\strong{Maintainer}: Andrei-Florian Stoica \email{andreistoica@foxmail.com} (\href{https://orcid.org/0000-0002-5253-0826}{ORCID})

}
\keyword{internal}
