% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{scUMAPMat}
\alias{scUMAPMat}
\title{Extracts the UMAP matrix from object.}
\usage{
scUMAPMat(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
A UMAP matrix.
}
\description{
This function extracts the UMAP matrix from a Seurat or
SingleCellExperiment object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='scLang')
sceObj <- qs::qread(scePath)
umapMat <- scUMAPMat(sceObj)

}
