% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sharedMarkers}
\alias{sharedMarkers}
\title{Join two marker data frames}
\usage{
sharedMarkers(df1, df2, joinCol = "avg_log2FC")
}
\arguments{
\item{df1}{A marker data frame.}

\item{df2}{A marker data frame.}

\item{joinCol}{Column based on which the marker data frames will be joined.}
}
\value{
A data frame with the shared markers as rows and the join column
from each of the two marker data frames as columns.
}
\description{
This function joins two marker data frames.
}
