% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimPlot}
\alias{dimPlot}
\title{Creates a dimensionality reduction plot}
\usage{
dimPlot(
  scObj,
  title = NULL,
  groupBy = NULL,
  dimred = "umap",
  dims = c(1, 2),
  legendTitle = "Group",
  noGroupsLegendLab = "Object",
  paletteFun = rainbow,
  alpha = 0.7
)
}
\arguments{
\item{title}{Plot title.}

\item{paletteFun}{Palette function. Must accept the number of colors as the
first argument and require no other arguments.}
}
\value{
A dimensionality reduction plot
}
\description{
This function creates a dimensionality reduction plot
}
