% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{dimPlot}
\alias{dimPlot}
\title{Creates a dimensionality reduction plot}
\usage{
dimPlot(
  scObj,
  title = NULL,
  groupBy = NULL,
  dimred = "umap",
  dims = c(1, 2),
  legendTitle = "Group",
  noGroupsLegendLab = "Object",
  paletteFun = rainbow,
  alpha = 0.7,
  legendPos = c("right", "top", "left", "bottom"),
  legendTextSize = 10,
  legendTitleSize = 10,
  axisTextSize = 12,
  axisTitleSize = 12,
  ...
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{title}{Plot title.}

\item{groupBy}{Grouping variable. Must exist in the metadata/coldata of
the single-cell expression object.}

\item{dimred}{Dimensionality reduction.}

\item{dims}{A numeric vector of size 2 representing the dimensions selected
for the plot.}

\item{legendTitle}{Legend title.}

\item{noGroupsLegendLab}{Legend label to be used when no grouping is
provided (\code{groupBy} is \code{NULL})}

\item{paletteFun}{Palette function. Must accept the number of colors as the
first argument and require no other arguments.}

\item{alpha}{Opaqueness level.}

\item{legendPos}{Legend position.}

\item{legendTextSize}{Legend text size.}

\item{legendTitleSize}{Legend title size.}

\item{axisTextSize}{Axis text size.}

\item{axisTitleSize}{Axis title size.}

\item{...}{Additional arguments passed to \code{centerTitle}.}
}
\value{
A dimensionality reduction plot.
}
\description{
This function creates a dimensionality reduction plot
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='scLang')
sceObj <- qs::qread(scePath)
dimPlot(sceObj, groupBy='Donor')

}
