% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{featurePlot}
\alias{featurePlot}
\title{Creates a dimensionality reduction plot}
\usage{
featurePlot(
  scObj,
  feature = rownames(scObj)[1],
  title = feature,
  dimred = "umap",
  dims = c(1, 2),
  legendTitle = feature,
  palette = wes_palette("Royal1")[c(3, 2)],
  alpha = 0.6,
  legendPos = c("right", "top", "left", "bottom"),
  legendTextSize = 10,
  legendTitleSize = 10,
  axisTextSize = 12,
  axisTitleSize = 12,
  ...
)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{feature}{A gene name or metadata column name.}

\item{title}{Plot title.}

\item{dimred}{Dimensionality reduction.}

\item{dims}{A numeric vector of size 2 representing the dimensions selected
for the plot.}

\item{legendTitle}{Legend title.}

\item{palette}{Color palette.}

\item{alpha}{Opaqueness level.}

\item{legendPos}{Legend position.}

\item{legendTextSize}{Legend text size.}

\item{legendTitleSize}{Legend title size.}

\item{axisTextSize}{Axis text size.}

\item{axisTitleSize}{Axis title size.}

\item{...}{Additional arguments passed to \code{centerTitle}.}
}
\value{
A feature plot.
}
\description{
This function creates a dimensionality reduction plot
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='scLang')
sceObj <- qs2::qs_read(scePath)
featurePlot(sceObj, 'Gene_0289')

}
