% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_and_percentages.R
\name{scColCounts}
\alias{scColCounts}
\title{Extract per-group counts from the column of a single-cell expression object}
\usage{
scColCounts(scObj, col = "orig.ident")
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}

\item{col}{Column as string.}
}
\value{
A frequency vector with the unique column values as names.
}
\description{
This function extracts per-group counts from the column of single-cell
expression object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='scLang')
sceObj <- qs::qread(scePath)
scColCounts(sceObj, 'Mutation_Status')

}
