% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_expression.R, R/generics.R
\name{scExpMat.default}
\alias{scExpMat.default}
\alias{scExpMat.Seurat}
\alias{scExpMat.SingleCellExperiment}
\alias{scExpMat.dgCMatrix}
\alias{scExpMat.matrix}
\alias{scExpMat}
\title{Extracts the expression matrix from object.}
\usage{
\method{scExpMat}{default}(
  scObj,
  dataType = c("data", "counts", "logcounts"),
  genes = NULL,
  densify = TRUE
)

\method{scExpMat}{Seurat}(
  scObj,
  dataType = c("data", "counts", "logcounts"),
  genes = NULL,
  densify = TRUE
)

\method{scExpMat}{SingleCellExperiment}(
  scObj,
  dataType = c("data", "counts", "logcounts"),
  genes = NULL,
  densify = TRUE
)

\method{scExpMat}{dgCMatrix}(
  scObj,
  dataType = c("data", "counts", "logcounts"),
  genes = NULL,
  densify = TRUE
)

\method{scExpMat}{matrix}(
  scObj,
  dataType = c("data", "counts", "logcounts"),
  genes = NULL,
  densify = TRUE
)

scExpMat(
  scObj,
  dataType = c("data", "counts", "logcounts"),
  genes = NULL,
  densify = TRUE
)
}
\arguments{
\item{scObj}{A \code{Seurat}, \code{SingleCellExperiment},
\code{dgCMatrix} or \code{matrix} object.}

\item{dataType}{Expression data type. Ignored if \code{scObj} is of class
\code{dgCMatrix} or \code{matrix}.}

\item{genes}{Selected genes. If \code{NULL}, all genes will be retained.}

\item{densify}{Whether to convert to dense matrix.}
}
\value{
An expression matrix.
}
\description{
This function extracts an expression matrix from a Seurat or
SingleCellExperiment object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs', package='scLang')
sceObj <- qs::qread(scePath)
mat <- scExpMat(sceObj, 'counts')

}
