% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_mat.R
\name{scPCAMat}
\alias{scPCAMat}
\title{Extracts the PCA matrix from object.}
\usage{
scPCAMat(scObj)
}
\arguments{
\item{scObj}{A \code{Seurat} or \code{SingleCellExperiment} object.}
}
\value{
A PCA matrix.
}
\description{
This function extracts the PCA matrix from a Seurat or
SingleCellExperiment object.
}
\examples{
scePath <- system.file('extdata', 'sceObj.qs2', package='scLang')
sceObj <- qs2::qs_read(scePath)
pcaMat <- scPCAMat(sceObj)

}
